requirements.txt
python-telegram-bot==21.4
yfinance==0.2.43
SQLAlchemy==2.0.25
apscheduler==3.10.4
